Commit final (debug + tweek)
15/ 20
The Great Loop (1 / 1)
Knowledge Representation (4 / 4)
Fitness (2 / 4)
Selection and cross-over (2 / 3)
Fine Tunning (3 / 4)
Does it works ? (3 / 4)
Advanced evolution (0 / 3)

+++ GeneticAlgorithmRobot/PlotDisplay.cs
@@ -15,9 +15,9 @@ namespace GeneticAlgorithmRobot
         Process gnuplot = new Process();
         StreamWriter gnupStWr;
 
-        public void Init()
+        public void Init(bool isLoaded)
         {
-            if (File.Exists(FILE_NAME))
+            if (/*!isLoaded &&*/ File.Exists(FILE_NAME))
                 File.Delete(FILE_NAME);
 
             gnuplot.StartInfo.FileName = @"C:\Program Files (x86)\gnuplot\bin\gnuplot.exe";
